Add checks for asg_deploy_name variable
The go-live and cleanup steps remove resources based on a regex pattern that
looks like this:
`{{asg_deploy_name}}.*`

If that variable is unintentionally overriden with an empty string outside the
role the pattern will match `.*` leading to asg, elb, and lc deletions/updates.

This check ensures that some value is set on the variable. If not it will fail
before doing damage.